consumers/requirements.txt
kafka-python==2.0.2
redis==5.0.8
requests==2.32.3
python-snappy==0.7.3
lz4==4.3.3
urllib3==2.0.7
dataclasses-json==0.6.1
pydantic==2.5.0
prometheus_client==0.17.1
